rename old sumcheck and new sumcheck (to delete old later)
