+++ polynomial/src/multilinear/evaluation_form.rs
@@ -43,7 +43,7 @@ impl<F: PrimeField> MultiLinearPolynomial<F> {
         assignments: &[F],
     ) -> Result<Self, &'static str> {
         // decided to go the consecutive partial evaluation route as opposed to the random access
-        // evaluation route because consecutive partial eval is all that's needed for sumcheck and
+        // evaluation route because consecutive partial eval is all that's needed for sumcheck_old and
         // gkr, and it seems random access partial evaluation will introduce additional cost (when
         // detecting duplicate assignments)
         let mut new_evaluations = self.evaluations.clone();
