+++ readme.md
@@ -4,7 +4,7 @@
 
 [Polynomial](/polynomial) - univariate polynomial, multilinear extension polynomial
 
-[Sumcheck](/sumcheck) - full sumcheck, paritial sumcheck (verifies to subclaim)
+[Sumcheck](/sumcheck_old) - full sumcheck, paritial sumcheck (verifies to subclaim)
 
 [GKR](/gkr) - gkr implementation
