+++ sumcheck/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "sumcheck"
+name = "sumcheck2"
 version = "0.1.0"
 edition = "2021"
 
@@ -8,5 +8,5 @@ edition = "2021"
 [dependencies]
 polynomial = { path = "../polynomial" }
 transcript = { path = "../transcript" }
-ark-serialize = { version = "0.4.1", features = ["derive"] }
 ark-ff = "0.4.1"
+ark-bls12-381 = "0.4.0"
