+++ sumcheck/src/lib.rs
@@ -1,264 +1,123 @@
-use crate::util::{partial_evaluation_points, skip_first_var_then_sum_over_boolean_hypercube};
+pub mod prover;
+pub mod verifier;
+
 use ark_ff::{BigInteger, PrimeField};
-use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
-use polynomial::univariate_poly::UnivariatePolynomial;
-use polynomial::Polynomial;
-use std::ops::Add;
-use transcript::Transcript;
 
-pub mod boolean_hypercube;
-pub mod util;
-
-#[derive(Debug, Clone)]
-pub struct SumcheckProof<F: PrimeField, P: Polynomial<F>> {
-    poly: P,
+#[derive(Debug)]
+/// Holds the round polys and the initial prover claimed sum for sumcheck_old
+pub struct SumcheckProof<F: PrimeField> {
     sum: F,
-    uni_polys: Vec<UnivariatePolynomial<F>>,
+    round_polys: Vec<Vec<F>>,
 }
 
-impl<F: PrimeField, P: Polynomial<F>> SumcheckProof<F, P> {
-    pub fn new(poly: P, sum: F) -> Self {
-        Self {
-            poly,
-            sum,
-            uni_polys: vec![],
-        }
-    }
-
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut result = vec![];
-        result.extend(self.poly.to_bytes());
-        result.extend(self.sum.into_bigint().to_bytes_be());
-        for poly in &self.uni_polys {
-            result.extend(poly.to_bytes());
-        }
-        result
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, CanonicalSerialize, CanonicalDeserialize)]
-/// Same as the sumcheck proof without the initial poly
-pub struct PartialSumcheckProof<F: PrimeField> {
-    pub sum: F,
-    uni_polys: Vec<UnivariatePolynomial<F>>,
-}
-
-impl<F: PrimeField, P: Polynomial<F>> From<SumcheckProof<F, P>> for PartialSumcheckProof<F> {
-    fn from(value: SumcheckProof<F, P>) -> Self {
-        Self {
-            sum: value.sum,
-            uni_polys: value.uni_polys,
-        }
-    }
-}
-
-impl<F: PrimeField> PartialSumcheckProof<F> {
-    pub fn to_bytes(&self) -> Vec<u8> {
-        let mut result = vec![];
-        result.extend(self.sum.into_bigint().to_bytes_be());
-        for poly in &self.uni_polys {
-            result.extend(poly.to_bytes());
-        }
-        result
-    }
-}
-
-#[derive(Debug, PartialEq)]
-/// Represents the result of a partial sumcheck proof verification
+/// Sometimes the verifier doesn't want to perform the final check
+/// in such cases, a subclaim is returned, this subclaim has all information
+/// needed to verify the last check
+/// sum = initial_poly(challenges)
 pub struct SubClaim<F: PrimeField> {
-    pub sum: F,
-    pub challenges: Vec<F>,
+    sum: F,
+    challenges: Vec<F>,
 }
 
-pub struct Sumcheck {}
-
-impl Sumcheck {
-    /// Generate a sum check proof given the poly and the claimed sum
-    pub fn prove<F: PrimeField, P: Polynomial<F>>(poly: P, sum: F) -> SumcheckProof<F, P>
-    where
-        for<'a> &'a P: Add<Output = Result<P, &'static str>>,
-    {
-        let mut transcript = Transcript::new();
-        transcript.append(poly.to_bytes().as_slice());
-
-        Self::prove_internal(poly, sum, &mut transcript).0
-    }
-
-    /// Generates a sumcheck proof that makes no statement about the initial poly
-    pub fn prove_partial<F: PrimeField, P: Polynomial<F>>(
-        poly: P,
-        sum: F,
-    ) -> (PartialSumcheckProof<F>, Vec<F>)
-    where
-        for<'a> &'a P: Add<Output = Result<P, &'static str>>,
-    {
-        let mut transcript = Transcript::new();
-        let (proof, challenges) = Self::prove_internal(poly, sum, &mut transcript);
-        (proof.into(), challenges)
-    }
-
-    fn prove_internal<F: PrimeField, P: Polynomial<F>>(
-        poly: P,
-        sum: F,
-        transcript: &mut Transcript,
-    ) -> (SumcheckProof<F, P>, Vec<F>)
-    where
-        for<'a> &'a P: Add<Output = Result<P, &'static str>>,
-    {
-        let mut uni_polys = vec![];
-        let mut challenges = vec![];
-
-        transcript.append(sum.into_bigint().to_bytes_be().as_slice());
-
-        for _ in 0..poly.n_vars() {
-            // partially evaluate the polynomial at the generated challenge points
-            let challenge_assignments =
-                partial_evaluation_points(poly.n_vars(), 0..challenges.len(), challenges.iter());
-            let challenge_poly = poly
-                .partial_evaluate(&challenge_assignments)
-                .unwrap()
-                .relabel();
-
-            let uni_poly = skip_first_var_then_sum_over_boolean_hypercube(challenge_poly)
-                .to_univariate()
-                .unwrap();
-            transcript.append(uni_poly.to_bytes().as_slice());
-            uni_polys.push(uni_poly);
-
-            // sample challenge
-            challenges.push(transcript.sample_field_element::<F>());
-        }
-
-        (
-            SumcheckProof {
-                poly,
-                sum,
-                uni_polys,
-            },
-            challenges,
-        )
-    }
-
-    /// Verify a sumcheck proof
-    pub fn verify<F: PrimeField, P: Polynomial<F>>(proof: SumcheckProof<F, P>) -> bool
-    where
-        for<'a> &'a P: Add<Output = Result<P, &'static str>>,
-    {
-        if proof.uni_polys.len() != proof.poly.n_vars() {
-            // number of round poly's should match total number of rounds
-            return false;
-        }
-
-        let mut transcript = Transcript::new();
-        // add poly to transcript
-        transcript.append(proof.poly.to_bytes().as_slice());
-
-        let initial_poly = proof.poly.clone();
-
-        if let Some(subclaim) = Self::verify_internal(proof.into(), &mut transcript) {
-            // final verifier check
-            // p_v(r_v) = p(r_1, r_2, ..., r_v)
-            let initial_poly_eval = initial_poly
-                .evaluate_slice(subclaim.challenges.as_slice())
-                .unwrap();
-            initial_poly_eval == subclaim.sum
-        } else {
-            false
-        }
-    }
-
-    /// Verify partial sumcheck proof
-    pub fn verify_partial<F: PrimeField>(proof: PartialSumcheckProof<F>) -> Option<SubClaim<F>> {
-        let mut transcript = Transcript::new();
-        Self::verify_internal(proof, &mut transcript)
-    }
-
-    fn verify_internal<F: PrimeField>(
-        proof: PartialSumcheckProof<F>,
-        transcript: &mut Transcript,
-    ) -> Option<SubClaim<F>> {
-        let mut challenges = vec![];
-
-        transcript.append(proof.sum.into_bigint().to_bytes_be().as_slice());
-
-        let mut claimed_sum = proof.sum;
-
-        for poly in proof.uni_polys {
-            // assert that p(0) + p(1) = sum
-            let p_0 = poly.evaluate(&F::zero());
-            let p_1 = poly.evaluate(&F::one());
-
-            if claimed_sum != (p_0 + p_1) {
-                return None;
-            }
-
-            // add poly to transcript
-            transcript.append(poly.to_bytes().as_slice());
-
-            // sample challenge and update claimed sum
-            let challenge = transcript.sample_field_element::<F>();
-            claimed_sum = poly.evaluate(&challenge);
-            challenges.push(challenge);
-        }
-
-        Some(SubClaim {
-            sum: claimed_sum,
-            challenges,
-        })
-    }
+/// Helper method for converting field elements to bytes
+fn field_elements_to_bytes<F: PrimeField>(field_elements: &[F]) -> Vec<u8> {
+    field_elements
+        .iter()
+        .map(|elem| elem.into_bigint().to_bytes_be())
+        .collect::<Vec<Vec<u8>>>()
+        .concat()
 }
 
 #[cfg(test)]
 mod tests {
-    use crate::Sumcheck;
-    use ark_ff::{Fp64, MontBackend, MontConfig};
+    use crate::prover::SumcheckProver;
+    use crate::verifier::SumcheckVerifier;
+    use ark_bls12_381::Fr;
     use polynomial::multilinear::coefficient_form::CoeffMultilinearPolynomial;
+    use polynomial::multilinear::evaluation_form::MultiLinearPolynomial;
+    use polynomial::product_poly::ProductPoly;
 
-    #[derive(MontConfig)]
-    #[modulus = "17"]
-    #[generator = "3"]
-    struct FqConfig;
-    type Fq = Fp64<MontBackend<FqConfig, 1>>;
-
-    fn p_2ab_3bc() -> CoeffMultilinearPolynomial<Fq> {
-        CoeffMultilinearPolynomial::new(
+    fn p_2ab_3bc() -> MultiLinearPolynomial<Fr> {
+        let evaluations = CoeffMultilinearPolynomial::new(
             3,
             vec![
-                (Fq::from(2), vec![true, true, false]),
-                (Fq::from(3), vec![false, true, true]),
+                (Fr::from(2), vec![true, true, false]),
+                (Fr::from(3), vec![false, true, true]),
             ],
         )
         .unwrap()
+        .to_evaluation_form();
+        MultiLinearPolynomial::new(3, evaluations).unwrap()
     }
 
     #[test]
-    fn test_sumcheck_protocol_correct_sum() {
+    fn test_sumcheck_correct_sum_multilinear() {
         // p = 2ab + 3bc
-        // sum over boolean hypercube = 10
         let p = p_2ab_3bc();
-        let sumcheck_proof = Sumcheck::prove(p, Fq::from(10));
-        assert!(Sumcheck::verify(sumcheck_proof));
+        let prod_poly = ProductPoly::new(vec![p]).unwrap();
+        let proof = SumcheckProver::<1, Fr>::prove(prod_poly.clone(), Fr::from(10)).unwrap();
+        let verification_result =
+            SumcheckVerifier::verify(prod_poly, proof).expect("proof is invalid");
+        assert!(verification_result);
     }
 
     #[test]
-    fn test_sumcheck_protocol_invalid_sum() {
+    fn test_correct_sum_multivariate_deg_2() {
+        // p = 2a^2b + 3ab
+        // decomposed into
+        // p1 = (2a + 0b + 3)
+        // p2 = (ab)
+        // p = p1 . p2
+
+        // p1 = (2a + 0b + 3)
+        let p1 = MultiLinearPolynomial::new(
+            2,
+            CoeffMultilinearPolynomial::new(
+                2,
+                vec![
+                    (Fr::from(2), vec![true, false]),
+                    (Fr::from(0), vec![false, true]),
+                    (Fr::from(3), vec![false, false]),
+                ],
+            )
+            .unwrap()
+            .to_evaluation_form(),
+        )
+        .unwrap();
+
+        // p2 = (ab)
+        let p2 = MultiLinearPolynomial::new(
+            2,
+            CoeffMultilinearPolynomial::new(2, vec![(Fr::from(1), vec![true, true])])
+                .unwrap()
+                .to_evaluation_form(),
+        )
+        .unwrap();
+
+        let p = ProductPoly::new(vec![p1, p2]).unwrap();
+
+        let proof = SumcheckProver::<2, Fr>::prove(p.clone(), Fr::from(5)).unwrap();
+        let verification_result = SumcheckVerifier::verify(p, proof).expect("proof is invalid");
+        assert!(verification_result);
+    }
+
+    #[test]
+    fn test_correct_sum_prove_partial() {
+        let p = p_2ab_3bc();
+        let prod_poly = ProductPoly::new(vec![p]).unwrap();
+        let (proof, _) =
+            SumcheckProver::<1, Fr>::prove_partial(prod_poly.clone(), Fr::from(10)).unwrap();
+        let subclaim = SumcheckVerifier::verify_partial(proof).expect("proof is invalid");
+        let expected_sum = prod_poly.evaluate(subclaim.challenges.as_slice()).unwrap();
+        assert_eq!(expected_sum, subclaim.sum);
+    }
+
+    #[test]
+    fn test_invalid_sum() {
         // p = 2ab + 3bc
-        // sum over boolean hypercube = 10
         let p = p_2ab_3bc();
-        let sumcheck_proof = Sumcheck::prove(p, Fq::from(200));
-        assert!(!Sumcheck::verify(sumcheck_proof));
-    }
-
-    #[test]
-    fn test_partial_sumcheck() {
-        // invalid sum
-        let partial_proof = Sumcheck::prove_partial(p_2ab_3bc(), Fq::from(200));
-        let verification_result = Sumcheck::verify_partial(partial_proof.0);
-        assert_eq!(verification_result, None);
-
-        // valid sum
-        let partial_proof = Sumcheck::prove_partial(p_2ab_3bc(), Fq::from(10));
-        let verification_result = Sumcheck::verify_partial(partial_proof.0);
-        assert!(verification_result.is_some());
+        let prod_poly = ProductPoly::new(vec![p]).unwrap();
+        let proof = SumcheckProver::<1, Fr>::prove(prod_poly.clone(), Fr::from(12)).unwrap();
+        assert!(SumcheckVerifier::verify(prod_poly, proof).is_err());
     }
 }
